mining working

+++ 3/blockchain/src/main/java/farias/blockchain/controller/dto/Start.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class Start {
-  public double difficulty;
+  public int difficulty;
 }
